requirements.txt
django~=3.1

../dj-kaos-dev-utils/
../dj-kaos-pkg-dev-utils/

django-address~=0.2.5
